+++ includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: ojessi <ojessi@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/06 20:37:53 by ojessi            #+#    #+#             */
-/*   Updated: 2019/05/13 20:34:46 by ojessi           ###   ########.fr       */
+/*   Updated: 2019/05/15 17:34:16 by ojessi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,7 @@ char	*ft_strjoin(char const *s1, char const *s2);
 char	*ft_free_join_rev(char *src, char *new);
 char	*ft_free_strsub(char const *s, unsigned int start, size_t len);
 char	*ft_free_strjoin_duo(char const *s1, char const *s2);
-char    *ft_itoa_base(int nbr, int base, int uppercase);
+char    *ft_itoa_base(long long int nbr, int base, int uppercase);
 char    *ft_strrev(char *s);
 char	*ft_atoa_binary_base(char *str, int base, int uppercase);
 char	*ft_free_strncpy(char *dest, const char *new, size_t n);
